requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==1.8.6
django-braces==1.11.0
django-embed-video==1.1.2
django-memcache-status==1.1
djangorestframework==3.2.3
idna==2.6
olefile==0.44
Pillow==4.3.0
python3-memcached==1.51
requests==2.18.4
urllib3==1.22
